Changed types

+++ Sales-Invoice-API/Domain/Entities/Invoice.cs
@@ -8,9 +8,9 @@ namespace Domain.Entities
 {
     public class Invoice
     {
-        public float Id { get; set; }
+        public long Id { get; set; }
         public DateTime Date { get; set; }
-        public float Number { get; set; }
+        public long Number { get; set; }
         public int Quantity { get; set; }
         public decimal NetValue { get; set; }
         public int Vat { get; set; }
@@ -21,7 +21,7 @@ namespace Domain.Entities
         {
         }
 
-        public Invoice(float id, DateTime date, float number, int quantity, decimal netValue,int vat, string description,
+        public Invoice(long id, DateTime date, long number, int quantity, decimal netValue,int vat, string description,
             string contractor)
         {
             Id = id;
